# frontend/src/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "RecoveredKeyShareDto" : {
        "type" : "object",
        "properties" : {
          "processPrivateKey" : {
            "type" : "string",
            "pattern" : "[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*"
          },
          "unrecoveredKeyShare" : {
            "type" : "string",
            "pattern" : "[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*"
          },
          "recoveredKeyShare" : {
            "type" : "string",
            "pattern" : "[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*"
          },
          "signedProcessInfo" : {
            "type" : "string",
            "pattern" : "[-_A-Za-z0-9]+=*\\.[-_A-Za-z0-9]*=*\\.[-_A-Za-z0-9]*=*"
          }
        }
      },
      "RecoveryProcessDto" : {
        "type" : "object",
        "required" : [ "id", "vaultId", "type", "processPublicKey", "recoveredKeyShares" ],
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "vaultId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "type" : {
            "$ref" : "#/components/schemas/Type2"
          },
          "details" : {
            "type" : "string"
          },
          "requiredKeyShares" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 2
          },
          "processPublicKey" : {
            "type" : "string"
          },
          "recoveredKeyShares" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/RecoveredKeyShareDto"
            },
            "minProperties" : 1
          }
        }
      },
...
      "Type2" : {
        "type" : "string",
        "enum" : [ "CHANGE_PERMISSIONS", "COUNCIL_CHANGE" ]
      },
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/emergency-access" : {
      "get" : {
        "summary" : "finds all recovery processes for vaults recoverable by the current user",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RecoveryProcessDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/emergency-access/{processId}" : {
      "put" : {
        "summary" : "starts a new recovery process",
        "parameters" : [ {
          "name" : "processId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecoveryProcessDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "process created"
          },
          "400" : {
            "description" : "invalid request, e.g. missing required fields"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/emergency-access/{processId}/abort" : {
      "delete" : {
        "summary" : "aborts an existing recovery process",
        "parameters" : [ {
          "name" : "processId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "existing recovery process aborted"
          },
          "404" : {
            "description" : "existing recovery process not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/emergency-access/{processId}/complete" : {
      "delete" : {
        "summary" : "completes an existing recovery process",
        "parameters" : [ {
          "name" : "processId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "existing recovery process completed"
          },
          "404" : {
            "description" : "existing recovery process not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/emergency-access/{processId}/recovered-key-shares" : {
      "post" : {
        "summary" : "adds recovered key share",
        "parameters" : [ {
          "name" : "processId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecoveredKeyShareDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "key share added"
          },
          "400" : {
            "description" : "invalid request, e.g. missing required fields"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/emergency-access/{vaultId}" : {
      "get" : {
        "summary" : "finds an existing recovery process",
        "parameters" : [ {
          "name" : "vaultId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RecoveryProcessDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "tags" : [ "Emergency Access Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "hub-backend API",
    "version" : "1.5.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}